Added SessionEndedEventHandler and supporting functionality in IGuestSessionController.

+++ Synthesis.Guest.Modules/Controllers/GuestSessionController.cs
@@ -220,23 +220,19 @@ namespace Synthesis.GuestService.Controllers
                 throw new ValidationFailedException(validationResult.Errors);
             }
 
-            try
+            var guestSession = await _guestSessionRepository.GetItemAsync(id);
+            if (guestSession == null)
             {
-                var guestSession = await _guestSessionRepository.GetItemAsync(x => x.Id == id);
-
-                await _projectGuestContextService.RemoveProjectGuestContextAsync(guestSession.SessionId);
-
-                await _guestSessionRepository.DeleteItemAsync(id);
-
-                _eventService.Publish(EventNames.GuestSessionDeleted, new GuidEvent(guestSession.Id));
-
-                await _projectLobbyStateController.RecalculateProjectLobbyStateAsync(guestSession.ProjectId);
+                return;
             }
-            catch (DocumentNotFoundException)
-            {
-                // We don't really care if it's not found.
-                // The resource not being there is what we wanted.
-            }
+
+            await _projectGuestContextService.RemoveProjectGuestContextAsync(guestSession.SessionId);
+
+            await _guestSessionRepository.DeleteItemAsync(id);
+
+            _eventService.Publish(EventNames.GuestSessionDeleted, new GuidEvent(guestSession.Id));
+
+            await _projectLobbyStateController.RecalculateProjectLobbyStateAsync(guestSession.ProjectId);
         }
 
         public async Task<GuestSession> GetGuestSessionAsync(Guid id)
@@ -258,12 +254,26 @@ namespace Synthesis.GuestService.Controllers
             throw new NotFoundException("GuestSession could not be found");
         }
 
+        public async Task<GuestSession> GetGuestSessionBySessionIdAsync(Guid sessionId)
+        {
+            // Since this method is not called by the GuestSessionModule, no validation is required.
+            // Also, we don't need to throw a NotFoundException when the guest session isn't found
+            // because the event handler that calls this will check for null.
+
+            var projectGuestContext = await _projectGuestContextService.GetProjectGuestContextAsync(sessionId.ToString());
+            if (projectGuestContext == null)
+            {
+                return null;
+            }
+
+            return await _guestSessionRepository.GetItemAsync(projectGuestContext.GuestSessionId);
+        }
+
         public async Task<IEnumerable<GuestSession>> GetMostRecentValidGuestSessionsByProjectIdAsync(Guid projectId)
         {
             var validationResult = _validatorLocator.Validate<ProjectIdValidator>(projectId);
             if (!validationResult.IsValid)
             {
-                _logger.Error("Failed to validate the projectId while attempting to retrieve GuestSession resources.");
                 throw new ValidationFailedException(validationResult.Errors);
             }
 
@@ -271,23 +281,25 @@ namespace Synthesis.GuestService.Controllers
 
             if (!projectResult.IsSuccess() || projectResult.Payload == null)
             {
-                var message = $"Failed to retrieve project: {projectId}. Message: {projectResult.ReasonPhrase}, StatusCode: {projectResult.ResponseCode}, ErrorResponse: {_synthesisObjectSerializer.SerializeToString(projectResult.ErrorResponse)}";
-                _logger.Error(message);
-                throw new NotFoundException(message);
+                throw new NotFoundException($"Failed to retrieve project: {projectId}. Message: {projectResult.ReasonPhrase}, StatusCode: {projectResult.ResponseCode}, ErrorResponse: {_synthesisObjectSerializer.SerializeToString(projectResult.ErrorResponse)}");
             }
 
             var project = projectResult.Payload;
 
-            var validGuestSessions = await _guestSessionRepository.GetItemsAsync(x => x.ProjectId == projectId &&
-                                                                                      x.ProjectAccessCode == project.GuestAccessCode &&
-                                                                                      x.GuestSessionState != GuestState.PromotedToProjectMember);
+            var validGuestSessions = await _guestSessionRepository.CreateItemQuery()
+                .Where(x => x.ProjectId == projectId &&
+                    x.ProjectAccessCode == project.GuestAccessCode &&
+                    x.GuestSessionState != GuestState.PromotedToProjectMember)
+                .ToListAsync();
 
             if (validGuestSessions == null || !validGuestSessions.Any())
             {
                 return new List<GuestSession>();
             }
 
-            var currentValidProjectGuestSessions = validGuestSessions.GroupBy(s => s.UserId).OrderBy(gs => gs.Key).Select(gs => gs.OrderByDescending(x => x.CreatedDateTime).FirstOrDefault());
+            var currentValidProjectGuestSessions = validGuestSessions.GroupBy(s => s.UserId)
+                .OrderBy(gs => gs.Key)
+                .Select(gs => gs.OrderByDescending(x => x.CreatedDateTime).FirstOrDefault());
 
             return currentValidProjectGuestSessions;
         }
@@ -317,10 +329,12 @@ namespace Synthesis.GuestService.Controllers
 
             var project = projectResult.Payload;
 
-            var validGuestSessions = await _guestSessionRepository.GetItemsAsync(x => x.ProjectId == projectId &&
-                x.UserId == userId &&
-                x.ProjectAccessCode == project.GuestAccessCode &&
-                x.GuestSessionState != GuestState.PromotedToProjectMember);
+            var validGuestSessions = await _guestSessionRepository.CreateItemQuery()
+                .Where(x => x.ProjectId == projectId &&
+                    x.UserId == userId &&
+                    x.ProjectAccessCode == project.GuestAccessCode &&
+                    x.GuestSessionState != GuestState.PromotedToProjectMember)
+                .ToListAsync();
 
             if (validGuestSessions == null || !validGuestSessions.Any())
             {
@@ -355,9 +369,11 @@ namespace Synthesis.GuestService.Controllers
 
             var project = projectResult.Payload;
 
-            var guestSessions = await _guestSessionRepository.GetItemsAsync(x => x.ProjectId == projectId &&
-                x.UserId == userId &&
-                x.ProjectAccessCode == project.GuestAccessCode);
+            var guestSessions = await _guestSessionRepository.CreateItemQuery()
+                .Where(x => x.ProjectId == projectId &&
+                    x.UserId == userId &&
+                    x.ProjectAccessCode == project.GuestAccessCode)
+                .ToListAsync();
 
             if (guestSessions == null || !guestSessions.Any())
             {
